Added files via upload
These are the main fundemental most recent files used in the analysis.

All To be summarized in 2 main map Files:
===============================
1- DIST_WMATA lines and stations 2
2- WMATA lines and stations with population density_copy

Note: These map files has the metro station layers already included from a map on Census Bureau, Esri, WMATA. Hence only the zones layer was added.

1 Data File
========
In addition to one main excel file : Nearest 200 Zones _ 86 stations _CLEANED. This file has the nearest 200 zones in a 100 miles range ranked in order to the closest zone to each station. This file will be the main file used in R to build the regression model about.